requirements.txt
aiofiles
aiohttp==3.8.1
pytest==7.1.2
PyYAML==6.0
gql==3.3.0
graphql-core==3.2.1
